Move utils files to folder
