+++ packages/bygger/src/components/FormMetaDataEditor/fields/BasicFields.tsx
@@ -1,7 +1,7 @@
 import { TextField } from '@navikt/ds-react';
 import { NavFormSettingsDiff, NavFormType, UsageContext } from '@navikt/skjemadigitalisering-shared-domain';
 import LabelWithDiff from '../LabelWithDiff';
-import { FormMetadataError, UpdateFormFunction } from '../utils';
+import { FormMetadataError, UpdateFormFunction } from '../utils/utils';
 import TemaKodeFields from './TemaKodeFields';
 import TestSkjemaFields from './TestSkjemaFields';
